adaptations in code to satisfy c11 standard

+++ libs/exec/discardmsg.c
@@ -15,7 +15,7 @@ void DiscardMsg(struct Message * msg)
     if (msg && msg->mn_ReceivePort)
     {
         if (((intptr_t)msg - (intptr_t)msg->mn_ReplyPort) == sizeof(uuid_t))
-            FreeVecPooled(msg->mn_ReceivePort->mp_MsgPool, (void *)msg - sizeof(uuid_t));
+            FreeVecPooled(msg->mn_ReceivePort->mp_MsgPool, (void *)((intptr_t)msg - sizeof(uuid_t)));
         else
             FreeVecPooled(msg->mn_ReceivePort->mp_MsgPool, msg);
     }
